+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change Log
 
-All notable changes to the "vscode-csky-cdk" extension will be documented in this file.
+All notable changes to the "csky-cdk-assistant" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 
 - Initial release
+- Discover and parse CDK projects directly without `cdk-make-assistant`.
+- Store project selection in `.vscode/csky-cdk.json`.